refactor: jest 周りのディレクトリ構成を微修正
