+++ src/__tests__/pages/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { render } from 'src/lib/test-utils'
 import HomePage from 'src/pages'
-import { posts } from 'src/__tests__/data/posts'
+import { posts } from '../data/posts'
+import { render } from '../utils'
 
 test('ホームページ', () => {
   const { container } = render(<HomePage posts={posts} />)
